Fix consumer ack timeouts in spark request consumer
We originally tried to fix this in c13b9f0114ad93ec21ec2a2ce42891c453c4b102.
The fix worked fine and we did not recompute requests again and again
in spark. However, we continued to get a consumer ack timeout daily during
the cron job runs.

It turns out that rabbitmq tries to send as many messages possible at
a time by default. All of these messages are delivered and marked as
unacked. The callback is called on first message. We ack it right away
and then execute the spark request. Once the request has finished the
method returns and the callback is executed on the next message. Now
if the first request took a considerable amount of time to execute,
we have already hit a timeout on the remaining messages before we
reach them. This is probably why we have been seeing the consumer
ack timeouts till now.

To fix this, we request rabbitmq to send only one message at a
time and ack the message after it has been processed. (If we ack'ed
right away, rabbitmq would send another message and we are back at
the above described problem.) The next message is not sent until
the current one has been ack'ed so the issue of timeout is mitigated.